后台管理页面新增登录记录查看；将文件上传路径改为了相对路径并重构保存方法

+++ src/main/java/com/tmyx/backend/controller/ArticleController.java
@@ -15,6 +15,7 @@ import org.springframework.web.multipart.MultipartFile;
 import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
+import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
@@ -36,7 +37,7 @@ public class ArticleController {
 
     // 获取所有文章
     @GetMapping("/all")
-    public Result getAllArticles() {
+    public Result getAllArticles(@RequestAttribute Integer userId) {
         List<Article> articles = articleMapper.findAll();
         return Result.success(articles);
     }
@@ -58,10 +59,11 @@ public class ArticleController {
     public Result preGenerateId() {
         // 生成文章id
         String articleId = UUID.randomUUID().toString().replace("-", "");
-
+        // 获取绝对路径
+        File baseDir = new File(resourcePath).getAbsoluteFile();
         String subPath = "/article/" + articleId + "/images/";
         File folder = new File(resourcePath + subPath);
-
+        // 如果目录不存在则创建
         if (!folder.exists()) {
             boolean success = folder.mkdirs();
             if (!success) {
@@ -73,24 +75,28 @@ public class ArticleController {
 
     // 上传图片
     @PostMapping("/uploadImage")
-    public Result uploadImage(@RequestParam("file") MultipartFile file,
+    public Result uploadImage(@RequestAttribute Integer userId,
+                              @RequestParam("file") MultipartFile file,
                               @RequestParam("articleId") String articleId) {
         if (file.isEmpty()) return Result.error("文件不能为空");
+        // 获取绝对路径
+        File baseDir = new File(resourcePath).getAbsoluteFile();
 
         try {
             // 确定文章的存储路径
-            String subPath = "article/" + articleId + "/images/";
-            File folder = new File(resourcePath + subPath);
+            String subPath = "/article/" + articleId + "/images/";
+            File folder = new File(baseDir + subPath);
+            // 如果目录不存在则创建
             if (!folder.exists()) {
                 folder.mkdirs();
             }
 
             String fileName = System.currentTimeMillis() + "_" + file.getOriginalFilename();
             File dest = new File(folder, fileName);
-            file.transferTo(dest);
+            file.transferTo(dest.getAbsoluteFile());
 
             // 拼接路径
-            String fileUrl = "/files/" + subPath + fileName;
+            String fileUrl = "/files" + subPath + fileName;
             return Result.success(fileUrl);
         } catch (IOException e) {
             e.printStackTrace();
@@ -100,22 +106,25 @@ public class ArticleController {
 
     // 保存文章
     @PostMapping("/save")
-    public Result saveArticle(@RequestBody Article article) {
+    public Result saveArticle(@RequestAttribute Integer userId, @RequestBody Article article) {
+        // 获取绝对路径
+        File baseDir = new File(resourcePath).getAbsoluteFile();
+        String absolutePath = baseDir.getAbsolutePath();
         // 确定文件存储相对路径以及文件名
         // article/{articleId}/index.md
-        String relativeFolder = "/article/" + article.getId() + "/";
-        String fileName = "index.md";
-        // 获取绝对路径
-        File folder = new File(resourcePath + relativeFolder);
+        String subPath = "/article/" + article.getId();
+        File folder = new File(baseDir + subPath);
+        // 如果目录不存在则创建
         if (!folder.exists()) folder.mkdirs();
         // 清除无用图片
         if (article.getContent() != null) {
-            FileUtil.cleanOrphanImages(article.getContent(), article.getId(), resourcePath);
+            FileUtil.cleanOrphanImages(article.getContent(), article.getId(), absolutePath);
         }
         // 将前端传来的content写入markdown文件中
         if (article.getContent() != null) {
-            try (FileWriter writer = new FileWriter(new File(folder, fileName))) {
-                writer.write(article.getContent());
+            try {
+                File mdFile = new File(folder, "index.md");
+                Files.writeString(mdFile.toPath(), article.getContent(), StandardCharsets.UTF_8);
             } catch (IOException e) {
                 e.printStackTrace();
                 return Result.error("内容文件保存失败");
@@ -123,10 +132,9 @@ public class ArticleController {
         } else {
             System.out.println("跳过文件写入");
         }
-
         // 组装文章信息
         article.setId(article.getId());
-        article.setContentUrl("/files" + relativeFolder + fileName);
+        article.setContentUrl("/files" + subPath + "/index.md");
         article.setUploadTime(LocalDateTime.now());
         // 检查文章是否存在
         Article existArticle = articleMapper.findById(article.getId());
@@ -140,15 +148,18 @@ public class ArticleController {
 
     // 编辑文章
     @GetMapping("/get/{id}")
-    public Result getById(@PathVariable String id) {
+    public Result getById(@RequestAttribute Integer userId, @PathVariable String id) {
         Article article = articleMapper.findById(id);
         if (article == null) return Result.error("文章不存在");
 
         String contentUrl = article.getContentUrl();
+        // 获取绝对路径
+        File baseDir = new File(resourcePath).getAbsoluteFile();
+        String absolutePath = baseDir.getAbsolutePath();
 
         if (contentUrl != null && !contentUrl.isEmpty()) {
             String relativePath = contentUrl.replace("/files/", "");
-            Path fullPath = Paths.get(resourcePath, relativePath);
+            Path fullPath = Paths.get(absolutePath, relativePath);
             File mdFile = fullPath.toFile();
 
             if (mdFile.exists()) {
@@ -167,7 +178,7 @@ public class ArticleController {
 
     // 修改文章标题
     @PutMapping("/updateTitle")
-    public Result updateTitle(@RequestBody ArticleUpdateDTO updateDTO) {
+    public Result updateTitle(@RequestAttribute Integer userId, @RequestBody ArticleUpdateDTO updateDTO) {
         int rows = articleMapper.updateTitle(updateDTO.getId(), updateDTO.getTitle());
 
         if (rows == 0) {
@@ -179,20 +190,27 @@ public class ArticleController {
 
     // 删除文章（删除整个目录）
     @DeleteMapping("/delete/{id}")
-    public Result deleteArticle(@PathVariable String id) {
+    public Result deleteArticle(@RequestAttribute Integer userId, @PathVariable String id) {
         // 根据文章id获取文章信息
         Article article = articleMapper.findById(id);
         if (article == null) return Result.error("文章不存在");
-        // 删除数据库记录
-        int rows = articleMapper.deleteById(id);
+
+        Path bassRoot = Paths.get(resourcePath).toAbsolutePath().normalize();
         // 获取文件路径
         String fileUrl = article.getContentUrl();
         // 去除映射路径的/files前缀，再去掉markdown文件的文件名，得到相对路径
         String relativeDirPath = fileUrl.replace("/files/", "").replace("index.md", "");
+        // 获取绝对路径
+        Path fullPath = bassRoot.resolve(relativeDirPath).normalize();
         // 拼接物理路径并删除
-        File articleFolder = new File(resourcePath + relativeDirPath);
-        FileUtil.deleteDirectory(articleFolder);
-
+        File articleFolder = fullPath.toFile();
+        if (articleFolder.exists() && articleFolder.isDirectory()) {
+            FileUtil.deleteDirectory(articleFolder);
+        } else {
+            System.out.println("目录不存在");
+        }
+        // 删除数据库记录
+        articleMapper.deleteById(id);
         return Result.success();
     }
 }
